+++ src/Features/Core/Portable/Completion/CompletionHelper.cs
@@ -495,14 +495,5 @@ namespace Microsoft.CodeAnalysis.Completion
                     initialTriggerKind == CompletionTriggerKind.Deletion;
             }
         }
-
-        public static async Task<SyntaxContext> CreateSyntaxContextWithExistingSpeculativeModelAsync(Document document, int position, CancellationToken cancellationToken)
-        {
-            Contract.ThrowIfFalse(document.SupportsSemanticModel, "Should only be called from C#/VB providers.");
-            var semanticModel = await document.ReuseExistingSpeculativeModelAsync(position, cancellationToken).ConfigureAwait(false);
-
-            var service = document.GetRequiredLanguageService<ISyntaxContextService>();
-            return service.CreateContext(document, semanticModel, position, cancellationToken);
-        }
     }
 }
